#1766 Moved tests covering the activatable logic into a separate package.
